requirements.txt
ml-collections
matplotlib
tensorflow-datasets
tensorflow-addons
tensorflow-text
pycocotools
scikit-image
einops
